Remove unused/duplicate file `build/lib/Gladiator.py`
The file `build/lib/Gladiator.py` was committed to the repo
(in violation of the `.gitignore` which excludes all `build/` files),
and there seems to be no reason to continue to have this file
in the repo as it is a near duplicate of `gladiator/Gladiator.py`
which seems to be the actual operative file in question.

This PR removes the unused/duplicate file to avoid confusion
for maintainers and contributors, and just to keep things tidy.