Add `debug_output` attribute to control debug printing.
Prior to this commit, the library would write a lot of stuff
to stdout. This is bad hygiene for a library that somebody
may want to use in their application - it should not log by default.

This commit fixes the issue by adding a `debug_output` attribute
that can be set to `True` if the user wants to see these logs.